PERF: Improved segmentation performance regarding less registrations
The usage of less registrations procedures improved total time of processing.
An additional checkbox for MNI152 registration was added, which allow the user to add the input data in MNI space.
CMakeList.txt was edited in order to automatically copy the resources data to the build directory (similar to SimpleFilters approach)