ENH: Improved general lesion segmentation and contrast enhancement
The global lesion segmentation process were revised, adding a more precise strategy to hyperintense lesions detection.
Modifications:
Logistic Contrast Enhancement: A BET histogram outlier removal was added in the algorithm, which improves the general lesion detection mainly for high lesion loads.
Weighted Enhancement image Filter: This module was reformulated which uses the entire contrast map in order to increase the input signal contrast. There is no longer used a threshold application on it, instead the full brain signal enhancement is applied. This guarantee the signal gaussianity, which was lacking in the previous implementation.
Python modules were revised and are functional.